mongoose userid,productid verify added, http header bug removal

+++ src/controllers/cartController.js
@@ -1,3 +1,4 @@
+const { default: mongoose } = require("mongoose");
 const CartItem = require("../models/cartModel");
 const User = require("../models/userModel");
 
@@ -7,10 +8,23 @@ const addToCart = async (req, res) => {
     const { userId } = req.params;
     const { productId, quantity } = req.body;
 
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res
+        .status(404)
+        .json({ message: "User not found with the provided ID" });
+    }
+    if (!mongoose.Types.ObjectId.isValid(productId)) {
+      return res
+        .status(404)
+        .json({ message: "Product not found with the provided ID" });
+    }
+
     // Find the User
     const user = await User.findById(userId);
 
-    if (!user) throw new Error("User not found with the provided ID");
+    if (!user) {
+      res.status().json({ message: "User not found with the provided ID" });
+    }
 
     // Check if the cart item already exist for this user
     const existingCartItem = await CartItem.findOne({ userId });
@@ -21,9 +35,9 @@ const addToCart = async (req, res) => {
       await existingCartItem.save();
     } else {
       //create new
-      await CartItem.create({ userId, productId, quantity });
+      return await CartItem.create({ userId, productId, quantity });
     }
-    res.send(201).json({ message: "Product added to cart" });
+    res.status(201).json({ message: "Product added to cart" });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -33,6 +47,12 @@ const addToCart = async (req, res) => {
 const getCartItems = async (req, res) => {
   try {
     const { userId } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res
+        .status(404)
+        .json({ message: "User not found with the provided ID" });
+    }
 
     const cartItems = await CartItem.find({ userId });
 
@@ -48,6 +68,18 @@ const updateCart = async (req, res) => {
     const { userId, productId } = req.params;
     const { quantity } = req.body;
 
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res
+        .status(404)
+        .json({ message: "User not found with the provided ID" });
+    }
+
+    if (!mongoose.Types.ObjectId.isValid(productId)) {
+      return res
+        .status(404)
+        .json({ message: "Product not found with the provided ID" });
+    }
+
     //find cart item
     const cartItem = await CartItem.findOneAndUpdate(
       { userId, productId },
@@ -55,7 +87,7 @@ const updateCart = async (req, res) => {
       { new: true }
     );
 
-    if (!cartItem) throw new Error("Cart Item Not Present");
+    if (!cartItem) res.status(404).json({ message: "Cart Item Not Present" });
 
     res.status(200).json({ message: "Cart Item Quantity Updated" });
   } catch (error) {
@@ -67,6 +99,18 @@ const updateCart = async (req, res) => {
 const removeFromCart = async (req, res) => {
   try {
     const { userId, productId } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res
+        .status(404)
+        .json({ message: "User not found with the provided ID" });
+    }
+
+    if (!mongoose.Types.ObjectId.isValid(productId)) {
+      return res
+        .status(404)
+        .json({ message: "Product not found with the provided ID" });
+    }
 
     // Find and delete cart item based on userId and productId
     const deleteCartItem = await CartItem.findOneAndDelete({
